requirements.txt
fastapi
uvicorn[standard]
requests
biopython